+++ webextension/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## *UNRELEASED*
 - Fix potential false positives on unsecured pages blocked by Firefox.
+- Remove SHA1 fingerprints comparison.
 
 ## 5.2.1 - 2019-02-01
 - Fix SSL Pinning on Check Server.